Implemented notification system for leave request workflow

+++ EnterpriseLeaveManagement.Application/Features/Dashboard/Responses/EmployeeDashboardResponse.cs
@@ -14,7 +14,7 @@ public class EmployeeDashboardResponse
 
     public int RejectedLeaveRequests { get; set; }
 
-    public int RemainingLeaveBalance { get; set; }
+    public decimal RemainingLeaveBalance { get; set; }
 
     public int UpcomingLeaves { get; set; }
 }
